Complete refactoring to address performance.
Found an issue where the entire grid was being searched for each neighbor for each cell. Refactored so that I can go directly to the object by name just by knowing the current position.

There is now an issue with Chrome and a maximum number of callbacks that may be due to lodash.